Working basic checks on resource, module, varibale names and exit codes.
