+++ app/__init__.py
@@ -1,0 +1,1 @@
+from app.terraform_checker import TerraformChecker
